replace Asciidoc with Markdown (fix #114)
in re #111 146394b24b:

When a @GitHub repository has a code of conduct file at
`/.github/code_of_conduct.md`, all goes according to plan. When,
however, an equivalent Asciidoc code of conduct is saved to
`/.github/code_of_conduct.adoc`, GitHub won’t see it.

What does that mean? There is a link to a repository’s “Code of conduct”
at the bottom of the “New issue” page, but only if the code of conduct
is saved as a Markdown file with a Markdown‑like filename extension ∴ no
Asciidoc codes of conduct for now.

Signed-off-by: Lucas Larson <LucasLarson@riseup.net>